Move SbtAccess to our own package below sbt to prevent accidental namespace clashes with other plugins / sbt
