+++ src/main/scala/sbt/dependencygraph/SbtAccess.scala
@@ -14,17 +14,13 @@
  *    limitations under the License.
  */
 
-package sbt
+package sbt.dependencygraph
 
-import internal._
-import Classpaths._
-import LibraryManagement._
-import internal.util.JLine
+import sbt.Defaults
 
 /** Accessors to private[sbt] symbols. */
 object SbtAccess {
   val unmanagedScalaInstanceOnly = Defaults.unmanagedScalaInstanceOnly
 
-  def getTerminalWidth: Int = JLine.usingTerminal(_.getWidth)
-
+  def getTerminalWidth: Int = sbt.internal.util.JLine.usingTerminal(_.getWidth)
 }
