--- src/sbt-test/sbt-dependency-graph/build.properties
@@ -1,1 +1,0 @@
-sbt.version=0.13.6
